requirements.linux.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==22.12.0
bleach
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.16
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docopt==0.6.2
drf-yasg==1.21.8
executing==2.2.0
fastjsonschema==2.21.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
ipython==8.12.3
jdatetime==3.8.2
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
loguru==0.7.3
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mssql-django==1.5
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
PyNaCl==1.5.0
pyodbc==5.2.0
pyperclip==1.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytse-client==0.18.0
pytz==2024.2
pyup==0.1.4
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tenacity==7.0.0
tinycss2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
win32_setctime==1.2.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
django-cors-headers==4.7.0